# api.yaml
openapi: 3.0.3
info:
  title: cli-mgt-svc
  version: "2.4"
  description: |
    REST API for user management, Network Element (NE) management,
    NE connection config, NETCONF config backup, command history, and bulk import.

    **Permission model:** Users have a single permission derived from `account_type`:
    `0` (SuperAdmin) and `1` (Admin) → `admin`; `2` (Normal) → `user`.
    Only `admin` users may call management APIs.

    **Authentication:** Most endpoints require `Authorization: Basic <jwt_token>`.
    Token obtained from `POST /aa/authenticate` — the `response_data` field already includes the `Basic ` prefix, paste it as-is.
servers:
  - url: http://localhost:3000
    description: Local
security: []
...
paths:
...
  /aa/authenticate:
    post:
      tags: [Authentication]
      summary: Login — get JWT
      operationId: authenticate
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              username: anhdt195
              password: "123"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Paste the `response_data` value from `POST /aa/authenticate` directly.
        The token already contains the `Basic ` prefix — paste it as-is.
        Example: `Basic eyJhbGciOiJIUzI1NiJ9...`

  # ── Reusable responses ───────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request body / params
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
...
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"

  # ── Schemas ──────────────────────────────────────────────────────────────────
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        response_data:
          type: string
          description: "Token string including 'Basic ' prefix (e.g. 'Basic eyJ...'). Use this value directly as the Authorization header."
          example: "Basic eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        response_code:
          type: string
          example: "200"
        system_type:
          type: string
          example: 5GC
...